ImageCreator allocated in a separate class
